+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed  
 
 - Service URL changed from https://sc-runoffmodelingservices.streamstats.usgs.gov/ to https://streamstats.usgs.gov/local/scrunoffservices
+- Changed PRF endpoint to take a list of PRF and Area inputs to calculate PRF value
 
 ### Deprecated